requirements.txt
torch>=2.0.0
transformers>=4.30.0
bitsandbytes>=0.39.0
accelerate
torchvision
pandas
numpy
matplotlib
Pillow
tqdm

